requirements.txt
Flask==1.1.1
requests==2.23.0
folium==0.10.1
Flask_WTF==0.14.3
python-dotenv==0.13.0
